+++ source/config.js
@@ -24,6 +24,7 @@ export var Ucum = {
    */
   dimLen_: 7,
 
+
   /**
    *  The characters used as valid operators in a UCUM unit expression,
    *  where '.' is for multiplication and '/' is for division.
@@ -63,14 +64,14 @@ export var Ucum = {
    * displaying messages on a web site; should be blank when output is
    * to a file.
    */
-  openEmphHTML_ : '<span class="emphSpan">',
+  openEmphHTML_ : ' <span class="emphSpan">',
 
   /**
    * Closing HTML used to emphasize portions of error messages.  Used when
    * displaying messages on a web site; should be blank when output is
    * to a file.
    */
-  closeEmphHTML_ : '</span>' ,
+  closeEmphHTML_ : '</span> ' ,
 
   /**
    * Message that is displayed when annotations are included in a unit